Add wrapper to ensure an image is loaded before parsing
