+++ jingle_sleigh/src/context/instruction_iterator.rs
@@ -66,7 +66,7 @@ mod test {
         let mut sleigh = ctx_builder
             .build(SLEIGH_ARCH)
             .unwrap();
-        sleigh.set_image(mov_eax_0.as_slice()).unwrap();
+        let sleigh = sleigh.set_image(mov_eax_0.as_slice()).unwrap();
         let instr = sleigh.read(0, 1).last().unwrap();
         assert_eq!(instr.length, 5);
         assert!(instr.disassembly.mnemonic.eq("MOV"));
@@ -87,7 +87,7 @@ mod test {
         let mut sleigh = ctx_builder
             .build(SLEIGH_ARCH)
             .unwrap();
-        sleigh.set_image(mov_eax_0.as_slice()).unwrap();
+        let sleigh = sleigh.set_image(mov_eax_0.as_slice()).unwrap();
         let instr: Vec<Instruction> = sleigh.read(0, 5).collect();
         assert_eq!(instr.len(), 4);
     }
