+++ jingle_sleigh/src/context/mod.rs
@@ -1,5 +1,6 @@
 mod builder;
 mod instruction_iterator;
+mod loaded;
 
 use crate::error::JingleSleighError;
 use crate::error::JingleSleighError::{LanguageSpecRead, SleighInitError};
@@ -21,6 +22,7 @@ use crate::VarNode;
 use cxx::{SharedPtr, UniquePtr};
 use std::fmt::{Debug, Formatter};
 use std::path::Path;
+use crate::context::loaded::LoadedSleighContext;
 
 pub struct SleighContext {
     ctx: UniquePtr<ContextFFI>,
@@ -124,46 +126,19 @@ impl SleighContext {
         &self.language_id
     }
 
-    pub fn set_image<T: Into<Image> + Clone>(&mut self, img: T) -> Result<(), JingleSleighError> {
+    pub fn set_image<T: Into<Image> + Clone>(mut self, img: T) -> Result<LoadedSleighContext, JingleSleighError> {
         self.image = Some(img.clone().into());
         self.ctx
             .pin_mut()
             .setImage(img.into())
-            .map_err(|_| ImageLoadError)
+            .map_err(|_| ImageLoadError)?;
+        Ok(LoadedSleighContext::new(self))
     }
 
     pub fn get_image(&self) -> Option<&Image> {
         self.image.as_ref()
     }
 
-    pub fn instruction_at(&self, offset: u64) -> Option<Instruction> {
-        let instr = self
-            .ctx
-            .get_one_instruction(offset)
-            .map(Instruction::from)
-            .ok()?;
-        if self
-            .image
-            .as_ref()?
-            .contains_range(offset..(offset + instr.length as u64))
-        {
-            Some(instr)
-        } else {
-            None
-        }
-    }
-
-    pub fn read(&self, offset: u64, max_instrs: usize) -> SleighContextInstructionIterator {
-        SleighContextInstructionIterator::new(self, offset, max_instrs, false)
-    }
-
-    pub fn read_until_branch(
-        &self,
-        offset: u64,
-        max_instrs: usize,
-    ) -> SleighContextInstructionIterator {
-        SleighContextInstructionIterator::new(self, offset, max_instrs, true)
-    }
 }
 
 #[cfg(test)]
