+++ jingle_sleigh/src/ffi/mod.rs
@@ -21,7 +21,7 @@ mod tests {
             SleighContextBuilder::load_ghidra_installation("/Applications/ghidra").unwrap();
 
         let mut sleigh = builder.build("x86:LE:64:default").unwrap();
-        sleigh.set_image(bytes.as_slice()).unwrap();
+        let sleigh = sleigh.set_image(bytes.as_slice()).unwrap();
         sleigh.instruction_at(0).unwrap();
     }
     #[test]
@@ -31,7 +31,7 @@ mod tests {
             SleighContextBuilder::load_ghidra_installation("/Applications/ghidra").unwrap();
 
         let mut sleigh = builder.build("x86:LE:64:default").unwrap();
-        sleigh.set_image(bytes.as_slice()).unwrap();
+        let sleigh = sleigh.set_image(bytes.as_slice()).unwrap();
         sleigh.instruction_at(0).unwrap();
     }
 
@@ -44,7 +44,7 @@ mod tests {
         let mut sleigh = ctx_builder
             .build(SLEIGH_ARCH)
             .unwrap();
-        sleigh.set_image(mov_eax_0.as_slice()).unwrap();
+        let mut sleigh = sleigh.set_image(mov_eax_0.as_slice()).unwrap();
         let instr1 = sleigh.instruction_at(0);
         sleigh.set_image(nops.as_slice()).unwrap();
         let instr2 = sleigh.instruction_at(0);
